+++ src/assets.rs
@@ -4,6 +4,7 @@ use axum::http::StatusCode;
 use axum::http::Uri;
 use axum::response::IntoResponse;
 use axum::response::Response;
+use reportify::ResultExt;
 
 use crate::FRONTEND;
 
@@ -22,7 +23,8 @@ pub(crate) async fn static_asset(uri: Uri) -> Response {
     let mut response = file.contents().to_vec().into_response();
     response.headers_mut().insert(
         header::CONTENT_TYPE,
-        HeaderValue::from_str(mime.as_ref()).expect("MIME type should be a valid header value"),
+        HeaderValue::from_str(mime.as_ref())
+            .assert_ok("a MIME type returned by mime_guess must be a valid HTTP header value"),
     );
     if file.path().file_name().and_then(|name| name.to_str()) == Some("index.html") {
         response